Added some high-level documentation; shifted and renamed scripts.
